requirements.txt
aiohttp>=2.2.5
alphabet_detector
beautifulsoup4
bleach
cachetools
configparser
currencyconverter
emoji==1.7.0
feedparser
future
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
gpytranslate
markdown2
multicolorcaptcha
pillow>=4.2.0
psutil
psycopg2-binary
pyYAML>=5.1.2
pyrate-limiter
python-telegram-bot==13.13
requests
spamwatch
speedtest-cli
sqlalchemy==1.4.40
telethon==1.16.4
pyrogram==2.0.41
ujson
pycryptodome
base58
pytimeparse
SibylSystem
telegraph
wget
ffmpeg-python
faker
aiofiles
TgCrypto
asyncio
python-arq
apscheduler
html2text
fontTools
opencv-python-headless
pykeyboard
random2
motor
pySmartDL
pymongo
pyromod
openapi-codec
pathlib
wheel-inspect
pickle-mixin
pymongo[srv]
dnspython
envparse
typing
bs4
humanize
redis~=4.2.2
uvicorn[standard]
fastapi
pykillerx
pymediainfo
httpx[http2]
